Question number 1
given a class diagram about gameObject which is a parent class where the damageableObject subclass is an abstract class that has an abstract method onKilled(), damageableObject also acts as a parent class that has player and monster subclasses. While the gameDemo class is used to run the programming that has been made.